Updated documentation and customizability of exchages
Apologies as documentation did not mention anything about database migrations.

Also added a check to see if database is connected. If connections fails, suggestions will be outputted.

You can now disable exchanges if exchanges if you don't want data from them. This is still a work in progress, and the implementation isn't the cleanest but it works.

Ideally I would want to ignore if an exchange is unable to get data for a certain cryptocurrency. Any help would be appreciated.